requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
torch==2.0.0
ultralytics==8.0.58